move extra functions to obs_util, update print statements
